CI: Fix false CI failures with package helper files
When touching helper files in category/package/files, they appear in
git diff and are thus parsed by tools/repoman_changed_files.sh. This
causes repoman to be run in category/package/files, which in turn causes
a repoman error, since the files subdirectory is not a valid directory
that contains ebuilds.

This commit fixes the CI pipeline by respecting only directores relative
to the repository root with a single / in their path (i.e. the package
diretories).

Signed-off-by: Adrian 'nex' Schollmeyer <adrian.schollmeyer@fem.tu-ilmenau.de>