media-video/srtrelay: Add missing dependency version requirement
A problem with net-libs/srt-1.4.1 was noticed where srtrelay was unable
to compile due to missing header files. Upgrading to net-libs/srt-1.4.2
fixed the problem. Thus, >=net-libs/srt-1.4.2 is being added as a
dependency.

Furthermore, net-libs/srt-1.4.3 introduces subslots. To be able to use
preserved-libs in the future, := is added to the dependency
specification.

Package-Manager: Portage-3.0.28, Repoman-3.0.3
Signed-off-by: Adrian Schollmeyer <nex+b-g-o@nexadn.de>