media-libs/gst-plugins-bad: Add package with fdkaac support
gst-plugins-bad from the main portage tree has no support for the
Fraunhofer AAC codec. However, adding support is as simple as adding
another USE flag and a conditional entry in RDEPEND.

This commit adds gst-plugins-bad as it is currently available in the
main portage tree, but with an additional fdk USE flag, which enables
support for the Fraunhofer AAC codec.

This patch could probably be upstreamed to the main portage tree, but is
in the fem-overlay for now, since it is required for the upcoming rC3
event.

Package-Manager: Portage-3.0.28, Repoman-3.0.3
Signed-off-by: Adrian Schollmeyer <nex+b-g-o@nexadn.de>