gitlab-ci: Explicitly run git fetch in repoman/pkgcheck jobs
Jobs working with the diff to origin/master require a git fetch to be
run before the actual check to work properly. If not run, the ref to
origin/master might be outdated and thus more packages are checked than
are actually affected by a merge request. This commit introduces git
fetch where necessary to fix this issue.