CI: Use custom docker image
The previous CI pipeline always built the required tools (repoman, git,
overlint, pkgcheck) and its dependencies on each CI run. This is very
inefficient and makes the pipeline take very long even for small
changes.

This commit replaces it with the newly built custom docker image,
providing a synced portage tree as well as the required tools already
installed. This reduces the time required to run the CI pipeline
drastically.