net-analyzer/fem-nagios-plugins: Update dependencies, refactor
This is a refactoring of the current 2.1.0 ebuild, since the old ebuild
uses way too many USE flags and had huge amounts of boilerplate code.
This commit introduces a PLUGIN_FLAG_MAP for plugins that should be
installed conditionally, allowing an easy mapping of USE flags to the
corresponding plugin names. Other plugins are in a PLUGIN_LIST so that
the installation only has to append the PLUGIN_LIST with the items from
PLUGIN_FLAG_MAP depending on whether the corresponding USE flags are
set and afterwards installing everythin in a for loop. This reduces
boilerplate code for checking each USE flag drastically and makes the
build easier to read.

All USE flags for plugins which don't introduces additional dependencies
are dropped. These plugins are now installed unconditionally, thus
reducing the amount of USE flags to be configured.

As all dependencies are only required at runtime, DEPEND is removed and
only RDEPEND is used. Furtheremore, a missing dependency for the hddtemp
plugin is added.

Closes: https://gitlab.fem-net.de/gentoo/fem-overlay/-/issues/98