net-dns/acme-dns: fix setting capabilities in init script
The previous init script didn't set the capabilities correctly, leading
to the daemon being unable to start due to missing permissions to listen
on udp/53. This commit adds a new revised init script which sets the
capabilities correctly (and which was successfully tested on a machine
running acme-dns as daemon).

Signed-off-by: Adrian 'nex' Schollmeyer <adrian.schollmeyer@fem.tu-ilmenau.de>